pull in JsonReader/JsonWriter class from AOSP
Summary:

jackson core is a big lib which has about 1k methods. It would be better if we can trim this down  and don't depend on it. My proposal is that we replace it with the streaming API android provides.

My short term goal is pulling all android json helper classes and repackge it into our lib. Use these classes to replace jackson. It should reduce the method footprint from 1k to 20.

Longterm goal is to depend on the system json helper class directly.

Task ID: #

Blame Rev:

Reviewers:

CC:

Test Plan:

1) no semantic change, the app still builds.